+++ BraintreePayPal/BTPayPalDriver.m
@@ -275,6 +275,7 @@ typedef NS_ENUM(NSUInteger, BTPayPalPaymentType) {
         }
 
         parameters[@"offer_paypal_credit"] = @(request.offerCredit);
+        parameters[@"offer_pay_later"] = @(request.offerPayLater);
 
         experienceProfile[@"no_shipping"] = @(!request.isShippingAddressRequired);
 
@@ -1014,6 +1015,12 @@ static NSString * const SFSafariViewControllerFinishedURL = @"sfsafariviewcontro
 
     if ((paymentType == BTPayPalPaymentTypeCheckout || paymentType == BTPayPalPaymentTypeBillingAgreement) && self.payPalRequest.offerCredit) {
         NSString *eventName = [NSString stringWithFormat:@"ios.%@.%@.credit.offered.%@", [self.class eventStringForPaymentType:paymentType], [self.class eventStringForRequestTarget:target], success ? @"started" : @"failed"];
+
+        [self.apiClient sendAnalyticsEvent:eventName];
+    }
+
+    if (paymentType == BTPayPalPaymentTypeCheckout && self.payPalRequest.offerPayLater) {
+        NSString *eventName = [NSString stringWithFormat:@"ios.%@.webswitch.paylater.offered.%@", [self.class eventStringForPaymentType:paymentType], success ? @"started" : @"failed"];
 
         [self.apiClient sendAnalyticsEvent:eventName];
     }